refactor(report): fix graph height and word-break
