+++ src/components/organisms/Histogram.tsx
@@ -72,6 +72,7 @@ export const Histogram = ({ data }: Props) => {
     data: histogramData,
     binField: "value",
     binWidth,
+    height: 200,
     colorField: "type",
     color: param => {
       const [binMin] = param.range as [number, number];
