+++ src/components/organisms/PieChart.tsx
@@ -16,6 +16,7 @@ export const PieChart = ({ data, colorMap }: Props) => {
     appendPadding: 10,
     data,
     angleField: "value",
+    height: 300,
     radius: 0.6,
     label: {
       type: "outer",
