+++ src/components/organisms/index.ts
@@ -9,4 +9,3 @@ export { default as LocalCouncilCard } from "./LocalCouncilCard";
 export { default as MapSelector } from "./MapSelector";
 export { default as MetroCouncilReport } from "./MetroCouncilReport";
 export { default as MetroCouncilCard } from "./MetroCouncilCard";
-export { default as NavigationCard } from "./NavigationCard";
